// src/app/features-modules/gestao-de-concurso/api/model/entidadeConcurso.ts
/**
 * Concurso.Api
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ConcursoCategoria } from './concursoCategoria';


export interface EntidadeConcurso { 
    id?: number;
    entidadeID?: number;
    nome?: string | null;
    descricao?: string | null;
    isMultiplas?: boolean;
    dataInicio?: string | null;
    dataFim?: string | null;
    concursoCategorias?: Array<ConcursoCategoria> | null;
}